v2.2.0-beta4
Minor documentation and code cleanup.

+++ Source/GrB_Descriptor_set.c
@@ -39,7 +39,7 @@ GrB_Info GrB_Descriptor_set     // set a parameter in a descriptor
                 return (GB_ERROR (GrB_INVALID_VALUE, (GB_LOG,
                     "invalid descriptor value [%d] for GrB_OUTP field;\n"
                     "must be GxB_DEFAULT [%d] or GrB_REPLACE [%d]",
-                    value, GxB_DEFAULT, GrB_REPLACE))) ;
+                    (int) value, (int) GxB_DEFAULT, (int) GrB_REPLACE))) ;
             }
             desc->out  = value ;
             break ;
@@ -51,7 +51,7 @@ GrB_Info GrB_Descriptor_set     // set a parameter in a descriptor
                 return (GB_ERROR (GrB_INVALID_VALUE, (GB_LOG,
                     "invalid descriptor value [%d] for GrB_MASK field;\n"
                     "must be GxB_DEFAULT [%d] or GrB_SCMP [%d]",
-                    value, GxB_DEFAULT, GrB_SCMP))) ;
+                    (int) value, (int) GxB_DEFAULT, (int) GrB_SCMP))) ;
             }
             desc->mask = value ;
             break ;
@@ -63,7 +63,7 @@ GrB_Info GrB_Descriptor_set     // set a parameter in a descriptor
                 return (GB_ERROR (GrB_INVALID_VALUE, (GB_LOG,
                     "invalid descriptor value [%d] for GrB_INP0 field;\n"
                     "must be GxB_DEFAULT [%d] or GrB_TRAN [%d]",
-                    value, GxB_DEFAULT, GrB_TRAN))) ;
+                    (int) value, (int) GxB_DEFAULT, (int) GrB_TRAN))) ;
             }
             desc->in0  = value ;
             break ;
@@ -75,7 +75,7 @@ GrB_Info GrB_Descriptor_set     // set a parameter in a descriptor
                 return (GB_ERROR (GrB_INVALID_VALUE, (GB_LOG,
                     "invalid descriptor value [%d] for GrB_INP1 field;\n"
                     "must be GxB_DEFAULT [%d] or GrB_TRAN [%d]",
-                    value, GxB_DEFAULT, GrB_TRAN))) ;
+                    (int) value, (int) GxB_DEFAULT, (int) GrB_TRAN))) ;
             }
             desc->in1  = value ;
             break ;
@@ -89,8 +89,8 @@ GrB_Info GrB_Descriptor_set     // set a parameter in a descriptor
                     "invalid descriptor value [%d] for GrB_AxB_METHOD field;\n"
                     "must be GxB_DEFAULT [%d], GxB_AxB_GUSTAVSON [%d]\n"
                     "GxB_AxB_HEAP [%d] or GxB_AxB_DOT [%d]",
-                    value, GxB_DEFAULT, GxB_AxB_GUSTAVSON, GxB_AxB_HEAP,
-                    GxB_AxB_DOT))) ;
+                    (int) value, (int) GxB_DEFAULT, (int) GxB_AxB_GUSTAVSON,
+                    (int) GxB_AxB_HEAP, (int) GxB_AxB_DOT))) ;
             }
             desc->axb  = value ;
             break ;
@@ -100,8 +100,9 @@ GrB_Info GrB_Descriptor_set     // set a parameter in a descriptor
             return (GB_ERROR (GrB_INVALID_VALUE, (GB_LOG,
                 "invalid descriptor field [%d], must be one of:\n"
                 "GrB_OUTP [%d], GrB_MASK [%d], GrB_INP0 [%d], GrB_INP1 [%d]"
-                "or GxB_AxB_METHOD [%d]", field,
-                GrB_OUTP, GrB_MASK, GrB_INP0, GrB_INP1, GxB_AxB_METHOD))) ;
+                "or GxB_AxB_METHOD [%d]", (int) field,
+                (int) GrB_OUTP, (int) GrB_MASK, (int) GrB_INP0, (int) GrB_INP1,
+                (int) GxB_AxB_METHOD))) ;
     }
 
     return (GrB_SUCCESS) ;
